Update de.js
Did the translation following the german core rule book of Star Wars - "Edge of the Empire" (german: "Am Rande des Imperiums").
I don't have the rule books of "Force and Destiny " and  "Age of Rebellion" yet (so the part of morality was new for me).
Also the german word for "Sig. Ability" is not clear at this point - means "sig ability" sth. like "designation" in this case?